+++ server/router/public-booking-router.go
@@ -31,6 +31,7 @@ type GetPublicBookableSpaceResponse struct {
 	LocationID     string `json:"locationId"`
 	LocationName   string `json:"locationName"`
 	RequireSubject bool   `json:"requireSubject"`
+	BookableDays   []int  `json:"bookableDays"`
 }
 
 type GetPublicBookableSpacesResponse struct {
@@ -156,6 +157,7 @@ func (router *PublicBookingRouter) getSpaces(w http.ResponseWriter, r *http.Requ
 			LocationID:     location.ID,
 			LocationName:   location.Name,
 			RequireSubject: space.RequireSubject,
+			BookableDays:   weekdaysFromString(location.BookableDays),
 		})
 	}
 	maxDaysInAdvance, _ := GetSettingsRepository().GetInt(orgID, SettingMaxDaysInAdvance.Name)
@@ -234,7 +236,7 @@ func (router *PublicBookingRouter) request(w http.ResponseWriter, r *http.Reques
 	}
 	language := normalizePublicBookingLanguage(m.Language)
 	name := strings.TrimSpace(m.Name)
-	if !IsValidHumanName(name) {
+	if !IsValidHumanName(name) || !ValidateEmail(m.Email) {
 		SendBadRequest(w)
 		return
 	}